Классы Courier и CourierRemoval разделены на класс работы с курьером и класс с проверками
