+++ src/main/java/order/OrderProperties.java
@@ -1,0 +1,4 @@
+package order;
+
+public class OrderProperties {
+}
